refactor:DogAnimationの座るモーション分離、型定義
